+++ src/controller/PacienteControllers/AntecedentesFamiliares-controller.js
@@ -1,26 +1,26 @@
-const { oMongoDB } = require("../database");
+const { oMongoDB } = require("../../database");
 const { ObjectId } = require("mongodb");
-const oFunctions = require("../helpers/functions");
+const oFunctions = require("../../helpers/functions");
 
-const getCasa = async (req, res) => {
+const getOneAntecedentesFamiliares = async (req, res) => {
   const { oID, oUserRol, oUserID } = req.body;
   let oCollection = await oMongoDB().collection("Paciente");
   let oQuery = { _id: new ObjectId(oID), isActive: 1 };
   let oResult = await oCollection.findOne(oQuery, {
     projection: {
-      Casa: 1,
+      AntecedentesFamiliares: 1,
     },
   });
   oFunctions.resetToken(oUserRol, oUserID);
   if (!oResult) res.send("NOT FOUND").status(404);
   else res.send(oResult).status(200);
 };
-const updateCasa = async (req, res) => {
-  const { oID, oUserRol, oUserID, Casa } = req.body;
+const updateAntecedentesFamiliares = async (req, res) => {
+  const { oID, oUserRol, oUserID, AntecedentesFamiliares } = req.body;
   const oQuery = { _id: new ObjectId(oID) };
   const oUpdate = {
     $set: {
-      Casa: Casa,
+      AntecedentesFamiliares: AntecedentesFamiliares,
     },
   };
   let oCollection = await oMongoDB().collection("Paciente");
@@ -29,6 +29,6 @@ const updateCasa = async (req, res) => {
   res.send(oResult).status(200);
 };
 module.exports = {
-  getCasa,
-  updateCasa,
+  getOneAntecedentesFamiliares,
+  updateAntecedentesFamiliares,
 };
