+++ src/controller/PacienteControllers/CirculoSocial-controller.js
@@ -1,26 +1,26 @@
-const { oMongoDB } = require("../database");
+const { oMongoDB } = require("../../database");
 const { ObjectId } = require("mongodb");
-const oFunctions = require("../helpers/functions");
+const oFunctions = require("../../helpers/functions");
 
-const getHabitos = async (req, res) => {
+const getOneCirculoSocial = async (req, res) => {
   const { oID, oUserRol, oUserID } = req.body;
   let oCollection = await oMongoDB().collection("Paciente");
   let oQuery = { _id: new ObjectId(oID), isActive: 1 };
   let oResult = await oCollection.findOne(oQuery, {
     projection: {
-      Habitos: 1,
+      CirculoSocial: 1,
     },
   });
   oFunctions.resetToken(oUserRol, oUserID);
   if (!oResult) res.send("NOT FOUND").status(404);
   else res.send(oResult).status(200);
 };
-const updateHabitos = async (req, res) => {
-  const { oID, oUserRol, oUserID, Habitos } = req.body;
+const updateCirculoSocial = async (req, res) => {
+  const { oID, oUserRol, oUserID, CirculoSocial } = req.body;
   const oQuery = { _id: new ObjectId(oID) };
   const oUpdate = {
     $set: {
-      Habitos: Habitos,
+      CirculoSocial: CirculoSocial,
     },
   };
   let oCollection = await oMongoDB().collection("Paciente");
@@ -29,6 +29,6 @@ const updateHabitos = async (req, res) => {
   res.send(oResult).status(200);
 };
 module.exports = {
-  getHabitos,
-  updateHabitos,
+  getOneCirculoSocial,
+  updateCirculoSocial,
 };
