+++ src/controller/PacienteControllers/EstadoMental-controller.js
@@ -1,26 +1,26 @@
-const { oMongoDB } = require("../database");
+const { oMongoDB } = require("../../database");
 const { ObjectId } = require("mongodb");
-const oFunctions = require("../helpers/functions");
+const oFunctions = require("../../helpers/functions");
 
-const getHistorialSexual = async (req, res) => {
+const getEstadoMental = async (req, res) => {
   const { oID, oUserRol, oUserID } = req.body;
   let oCollection = await oMongoDB().collection("Paciente");
   let oQuery = { _id: new ObjectId(oID), isActive: 1 };
   let oResult = await oCollection.findOne(oQuery, {
     projection: {
-      HistorialSexual: 1,
+      EstadoMental: 1,
     },
   });
   oFunctions.resetToken(oUserRol, oUserID);
   if (!oResult) res.send("NOT FOUND").status(404);
   else res.send(oResult).status(200);
 };
-const updateHistorialSexual = async (req, res) => {
-  const { oID, oUserRol, oUserID, HistorialSexual } = req.body;
+const updateEstadoMental = async (req, res) => {
+  const { oID, oUserRol, oUserID, EstadoMental } = req.body;
   const oQuery = { _id: new ObjectId(oID) };
   const oUpdate = {
     $set: {
-      HistorialSexual: HistorialSexual,
+      EstadoMental: EstadoMental,
     },
   };
   let oCollection = await oMongoDB().collection("Paciente");
@@ -29,6 +29,6 @@ const updateHistorialSexual = async (req, res) => {
   res.send(oResult).status(200);
 };
 module.exports = {
-  getHistorialSexual,
-  updateHistorialSexual,
+  getEstadoMental,
+  updateEstadoMental,
 };
