+++ src/controller/PacienteControllers/Problematica-controller.js
@@ -1,26 +1,26 @@
-const { oMongoDB } = require("../database");
+const { oMongoDB } = require("../../database");
 const { ObjectId } = require("mongodb");
-const oFunctions = require("../helpers/functions");
+const oFunctions = require("../../helpers/functions");
 
-const getOneCirculoSocial = async (req, res) => {
+const getProblematica = async (req, res) => {
   const { oID, oUserRol, oUserID } = req.body;
   let oCollection = await oMongoDB().collection("Paciente");
   let oQuery = { _id: new ObjectId(oID), isActive: 1 };
   let oResult = await oCollection.findOne(oQuery, {
     projection: {
-      CirculoSocial: 1,
+      Problematica: 1,
     },
   });
   oFunctions.resetToken(oUserRol, oUserID);
   if (!oResult) res.send("NOT FOUND").status(404);
   else res.send(oResult).status(200);
 };
-const updateCirculoSocial = async (req, res) => {
-  const { oID, oUserRol, oUserID, CirculoSocial } = req.body;
+const updateProblematica = async (req, res) => {
+  const { oID, oUserRol, oUserID, Problematica } = req.body;
   const oQuery = { _id: new ObjectId(oID) };
   const oUpdate = {
     $set: {
-      CirculoSocial: CirculoSocial,
+      Problematica: Problematica,
     },
   };
   let oCollection = await oMongoDB().collection("Paciente");
@@ -29,6 +29,6 @@ const updateCirculoSocial = async (req, res) => {
   res.send(oResult).status(200);
 };
 module.exports = {
-  getOneCirculoSocial,
-  updateCirculoSocial,
+  getProblematica,
+  updateProblematica,
 };
