+++ src/models/tablasVacias.js
@@ -69,14 +69,16 @@ const oPaciente = {
     ingresoFamiliar: "",
     serviciosHogar: [""],
   },
-  Familiar: {
-    nombre: "",
-    apellido: "",
-    parentesco: "",
-    ocupacion: "",
-    fechaNacimiento: "",
-    sustancia: [""],
-  },
+  Familiar: [
+    {
+      nombre: "",
+      apellido: "",
+      parentesco: "",
+      ocupacion: "",
+      fechaNacimiento: "",
+      sustancia: [""],
+    },
+  ],
   Sustancia: [""],
   nivelEscolar: "",
   religion: "",
