+++ src/routes/PacienteRoutes/AntecedentesClinicos-route.js
@@ -1,7 +1,7 @@
 const express = require("express");
 const router = express.Router();
-const oAntecedentesClinicos = require("../controller/AntecedentesClinicos-controller");
-const oMiddleware = require("../helpers/auth");
+const oAntecedentesClinicos = require("../../controller/PacienteControllers/AntecedentesClinicos-controller");
+const oMiddleware = require("../../../helpers/auth");
 
 router.post(
   "/byid",
