+++ src/routes/PacienteRoutes/Casa-route.js
@@ -1,7 +1,7 @@
 const express = require("express");
 const router = express.Router();
-const oCasa = require("../controller/Casa-controller");
-const oMiddleware = require("../helpers/auth");
+const oCasa = require("../../controller/PacienteControllers/Casa-controller");
+const oMiddleware = require("../../helpers/auth");
 
 router.post("/byid", oMiddleware.adminAuth, oCasa.getCasa);
 router.patch("/update", oMiddleware.adminAuth, oCasa.updateCasa);
