+++ src/routes/PacienteRoutes/EstadoMental-route.js
@@ -1,7 +1,7 @@
 const express = require("express");
 const router = express.Router();
-const oEstadoMental = require("../controller/EstadoMental-controller");
-const oMiddleware = require("../helpers/auth");
+const oEstadoMental = require("../../controller/PacienteControllers/EstadoMental-controller");
+const oMiddleware = require("../../helpers/auth");
 
 router.post("/byid", oMiddleware.adminAuth, oEstadoMental.getEstadoMental);
 router.patch(
