+++ src/routes/PacienteRoutes/Habitos-route.js
@@ -1,7 +1,7 @@
 const express = require("express");
 const router = express.Router();
-const oHabitos = require("../controller/Habitos-controller");
-const oMiddleware = require("../helpers/auth");
+const oHabitos = require("../../controller/PacienteControllers/Habitos-controller");
+const oMiddleware = require("../../helpers/auth");
 
 router.post("/byid", oMiddleware.adminAuth, oHabitos.getHabitos);
 router.patch("/update", oMiddleware.adminAuth, oHabitos.updateHabitos);
